requirement.txt
label-studio==1.6.0
label-studio-converter==0.0.44
label-studio-ml==1.0.8
label-studio-tools==0.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.2
mmcv-full==1.5.1
mmdet==2.24.1
numpy==1.22.4
opencv-python==4.5.5.64
Pillow==9.0.1
pip==22.3.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
zipp==3.8.0
